added: detection of edge points and center points of rectangle added: networking support under RobotVisionNetworking class
fixed: no more warnings (signed/unsigned mismatch and such)

This code compiles but has yet to be tested...

+++ trunk/OpenCV Robot Vision/main.cpp
@@ -106,7 +106,7 @@ void main()
 		if (frames%15 == 0)
 		{
 			end = time(NULL);
-			float diffTime = difftime(end, start);
+			float diffTime = (float)difftime(end, start);
 
 			// output framerate
 			cout<< "FPS: " << 1/(diffTime) << endl;
